Move unit tests into a subfolder
