+++ setup.cfg
@@ -90,7 +90,7 @@ source =
     tavern
 
 [tool:pytest]
-testpaths=tavern tests
+testpaths=tavern tests/unit
 addopts =
     --cov-report term-missing
     --cov tavern
